Copied Jorg Scene, prevented double jumping after falling, added jump grace period
Copied the Jorg Scene to the Movement branch.

The player can no longer jump a second time after falling off a platform. This was done to prevent the player from covering a greater distance.

The jump grace period is a customizable small window of time in which the player can still perform double jumps after falling off a platform.