Melee Enemy (Working with no damage or knockback))
My script allows for enemies to dash toward the Player's last seen position at the rate of melee delay. You can adjust speed. However, you can't adjust distance as it will only go as far as the player's distance. This isn't the intention, but this will have to do for now. Damage and knockback have not been made applied yet, but will in the future. If by some inking of cosmic bad luck I don't finish this, feel free to use this version as a way to finish this version. I have doubts you will have to though :)