Added new checkpoint system and some minor updates
Added checkpoint prefab for ease of use. To use it, you will need an object with the PlayerSpawn tag. This object will teleport to the 'Spawn Pos' object in the checkpoint whenever the player touches the checkpoint. Feel free to move around the 'Spawn Pos' object anywhere near the checkpoint, the player will spawn in that position.

Added option to the Damage script that respawns the player to the current spawn point when damaged (used in the water from Jorg Plains).

Debug option added that allows the player to be spawned where the scene editor camera is.

Debug sprint removed.

Jorg Scene cleanup.

Resuscitated Oiiai Cat.

Few adjustments to Jorg Plains.

Commented out some debug logs.