Non-Audio Settings and Updated Player UI (For Peer Review)
this is just to showcase the functionality, so no fancy settings menu because chris is working on it actively. the sliders are in the pause menu. one changes the FOV, the other changes the camera sensitivity, and the last changes the screen brightness (a simulated one)

the things added to the UI are SettingsData, Player Dash Cooldown, and Brightness Overlay. The script changes are up to date with the most recent commit in safety